CSSInlineEditor refactoring: * Move CSSInlineEditor class into separate file and rename to MultiRangeInlineEditor * Change NEXT_CSS_RULE and PREV_CSS_RULE to QUICK_EDIT_NEXT_MATCH/QUICK_EDIT_PREV_MATCH
There are no functional changes and the only UI change is with the "Next CSS Rule" and "Previous CSS Rule" menu items (they are now called "Next Match"/"Previous Match")

This refactoring allows other types of quick editors to be written as extensions.
