+++ src/command/Commands.js
@@ -63,10 +63,10 @@ define(function (require, exports, module) {
     exports.VIEW_HIDE_SIDEBAR  = "view.hideSidebar";
     
     // Navigate
-    exports.NAVIGATE_QUICK_OPEN = "navigate.quickOpen";
-    exports.SHOW_INLINE_EDITOR  = "navigate.showInlineEditor";
-    exports.NEXT_CSS_RULE       = "navigate.nextCssRule";
-    exports.PREVIOUS_CSS_RULE   = "navigate.previousCssRule";
+    exports.NAVIGATE_QUICK_OPEN     = "navigate.quickOpen";
+    exports.SHOW_INLINE_EDITOR      = "navigate.showInlineEditor";
+    exports.QUICK_EDIT_NEXT_MATCH   = "navigate.nextMatch";
+    exports.QUICK_EDIT_PREV_MATCH   = "navigate.previousMatch";
 
     exports.DEBUG_REFRESH_WINDOW = "debug.refreshWindow"; // string must MATCH string in native code (brackets_extensions)
     exports.DEBUG_SHOW_DEVELOPER_TOOLS = "debug.showDeveloperTools";
