+++ src/index.html
@@ -149,8 +149,8 @@
                             <li><a href="#" id="menu-navigate-go-to-symbol">Go to Symbol</a></li>
                             <li><hr class="divider"></li>
                             <li><a href="#" id="menu-navigate-quick-edit">Quick Edit</a></li>
-                            <li><a href="#" id="menu-navigate-previous-css-rule">Previous CSS Rule</a></li>
-                            <li><a href="#" id="menu-navigate-next-css-rule">Next CSS Rule</a></li>
+                            <li><a href="#" id="menu-navigate-previous-match">Previous Match</a></li>
+                            <li><a href="#" id="menu-navigate-next-match">Next Match</a></li>
                         </ul>
                     </li>
